chore(api): move solution-service-qrocm-ind.js to evaluation scope
