+++ api/tests/evaluation/unit/domain/services/solution/solution-service-qrocm-ind_test.js
@@ -1,7 +1,7 @@
-import * as service from '../../../../lib/domain/services/solution-service-qrocm-ind.js';
-import { YamlParsingError } from '../../../../src/shared/domain/errors.js';
-import { AnswerStatus } from '../../../../src/shared/domain/models/AnswerStatus.js';
-import { catchErr, expect } from '../../../test-helper.js';
+import * as service from '../../../../../../src/evaluation/domain/services/solution/solution-service-qrocm-ind.js';
+import { YamlParsingError } from '../../../../../../src/shared/domain/errors.js';
+import { AnswerStatus } from '../../../../../../src/shared/domain/models/AnswerStatus.js';
+import { catchErr, expect } from '../../../../../test-helper.js';
 
 const ANSWER_OK = AnswerStatus.OK;
 const ANSWER_KO = AnswerStatus.KO;
